+++ src/FREffects.hpp
@@ -3,11 +3,7 @@
 class FREffects {
 public:
     static CCSprite* grdInfinity();
-    static std::pair<CCSprite*, CCSprite*> infinityBackgrounds(CCLayer*, bool);
-    static CCSprite* mythicalBackground(CCLayer*);
-    static CCSprite* legendaryBackground(CCLayer*);
-    static CCParticleSystemQuad* legendaryParticles();
-    static CCParticleSystemQuad* mythicalParticles();
-    static CCParticleSystemQuad* infiniteParticles1(bool);
-    static CCParticleSystemQuad* infiniteParticles2();
+    static std::pair<CCSprite*, CCSprite*> infinityBackgrounds(bool);
+    static CCSprite* mythicalBackground();
+    static CCSprite* legendaryBackground();
 };